Extract objects
